Remove logic related to race condition prevention
resetprop already fundamentally breaks the existing system property
contract, which is "single writer, multiple reader", so it doesn't
make sense to spend any effort trying to workaround race conditions.

First of all, since we need to use the same code to handle pre API 30
devices, we cannot use the dirty backup area mechanism. After removing
that, we cannot rely on the dirty bit of the serial to function
correctly, so we skip that too.

Finally, we revert property read back to the old wait implementation
so that it works on all API levels.